refactor(LineBoard): 重複コードを共通化
LineBoardコンポーネント群で重複していたコード（約465行）を共通化しました。

主な変更:
- 共通ディレクトリ構造を作成 (src/components/LineBoard/shared/)
- useBarStylesフックを共通化
- LineDot、StationName、EmptyStationNameCellコンポーネントを共通化
- 共通スタイルを抽出 (commonStyles.ts)
- LineBoardEast、LineBoardSaikyo、LineBoardJRKyushu、LineBoardToeiで共通コンポーネントを使用

これにより、コードの保守性が向上し、将来の変更がより容易になります。
